rename LocationUpdateService to LocationUpdateManager
